Corrected player script
Corrected player script so the guns will fire when the player picks up a weapon but will not fire when the player does not have a weapon

+++ DRS_FPS_Prototype/Assets/Scripts/WeaponStats.cs
@@ -9,7 +9,7 @@ public class WeaponStats : ScriptableObject
     public GameObject Model;
     public int Damage;
     public int EffectiveRange;
-    public int RateOfFire;
+    public float RateOfFire;
     public int CurrentAmmo;
     public int MaxAmmo;
